Move components, defaults and helpers to their own files
