+++ src/index.js
@@ -1,0 +1,3 @@
+import ImageZoom from './ImageZoom'
+
+export default ImageZoom
